requirements.txt
fastapi
pydantic
python-dotenv
mysql-connector-python
langchain
langchain_google_genai
google-generativeai